Just some test files to get things going
These files help setup a vagrant vm that is running mongodb. For now, there are some post-install steps that need to be run on the vm: cd /vagrant && sudo ./post-install. That script also mentions a line in /etc/mongod.conf that needs to be commented out to allow mongo to bind to any interface