Reduce circle size in compass and clock
This allows the text in the clock to be adjusted to match the positionining (padding) of other components. We adjust both the compass and clock so they will be consistent sizes. This will also make room below the compass if we wish to show a text version of the current heading.