Adjust thruster layout
This makes each thruster a bit larger which required some nudging of their positions so they didn’t overlap. The skeleton color is now a neutral grey so as not to distract the eye. The red and green colors are now fully opaque making them a bit easier to see