Break out panels to left and right sides
I’m not very good with CSS layout, but this is mostly working. This centers the video between the left and right panel views. The left panel sticks to the left of the screen. The right panel sticks to the left of the video; however, the 3D view grows in width and I’d rather have the right panel stick to the rigth of the window.