Update thruster controller per feedback from last test
This removes the per platform mapping of controller axes. This code will always run on Raspbian so the mapping has to be done on the client.

This fixes the thruster interpolator names to match the physical layout on the submarine.

This reverse the thrustes when applying pitch to the submarine.

This remove thrust from the center vertical thruster (the one on the tail) when applying the ascent/descent axes.