Start work on visual controller/thruster calibration
This implements the Interpolator class from our Python implementation. We capture the logic for sensitivity curves as well. We plot graphs of a given interpolator, the current settings for sensitivity, and the combined effect of the sensitivity being applied to the interpolated values. Sliders are available to alter the sensitivity’s interpolation and power values with live updating in the graphs.