Add interactive thruster graphic
This allows the user to see, graphically, which thruster is being edited. This is in addition to the active thruster label that already exists.

The user already has the option to move through the thrusters via the keyboard; however, this adds the ability to click on the desired thruster as an alternate means for selecting the thruster to edit.