Refine Handle support
This now constrains movement to a bbox, when one is provided. This required an update to how the drag region is updated.

This also adjusts which elements get which mouse events. Previously all events were on the drag region which could have made it difficult to select points when they were close to other points.