Changed control value update
A rough attempt at exclusively storing the controller values that have changed status from the previous values in the form of a multidimensional array of current and former axis/button information. I am not sure how sound my logic here is in terms of what we are trying to achieve, but this is how I interpreted what the code should be doing in order to maintain efficient data traffic and avoid sending useless values. Please let me know if I have the right thought process here.