WIP: begin tracking gamepad changes
This code still needs more work, but the idea is that we now have a class that will track the last known values of a given game controller. Each time we update the current state, events will be fired for each change. Changes are sent to the ThrusterClient which is reponsible for sending the change to the thruster server